Began refactor for ulps_eq
